third_party/blink/web_tests/platform/mac/plugins/focus-change-2-change-focus-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(true)
i1focus
i1blur
pfocus-change-2
pblur
i2focus
i2blur
i3focus
PASS successfullyParsed is true

TEST COMPLETE


Test changing focus inside plugin onfocus event

   
